+++ Application/Commands/CartCommands/AddProduct/AddProductInCartCommandHandler.cs
@@ -2,6 +2,7 @@
 using AutoMapper;
 using Data;
 using MediatR;
+using Microsoft.Extensions.Configuration;
 using Persistence.Repository;
 using System;
 using System.Collections.Generic;
@@ -9,27 +10,28 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Application.Commands.CartCommands.UpdateProduct
+namespace Application.Commands.CartCommands.AddProduct
 {
-    public class UpdateProductCommandHandler : IRequestHandler<UpdateProductCommand, CartModel>
+    public class AddProductInCartCommandHandler : IRequestHandler<AddProductInCartCommand, CartModel>
     {
         private readonly IRepository<Cart> _cartRepository;
+        private readonly IRepository<StorageItem> _storageItemRepository;
         private readonly IMapper _mapper;
 
-        public UpdateProductCommandHandler(IMapper mapper, IRepository<Cart> cartRepository)
+        public AddProductInCartCommandHandler(IMapper mapper, IRepository<Cart> cartRepository,
+            IRepository<StorageItem> storageItemRepository)
         {
             _cartRepository = cartRepository;
+            _storageItemRepository = storageItemRepository;
             _mapper = mapper;
         }
-
-        public async Task<CartModel> Handle(UpdateProductCommand request, CancellationToken cancellationToken)
+        public async Task<CartModel> Handle(AddProductInCartCommand request, CancellationToken cancellationToken)
         {
             if (request == null) throw new ArgumentException();
 
             var cart = await _cartRepository.GetById(request.CartId);
 
-            var product = cart.Items.FirstOrDefault(item => item.Product.Id == request.ProductId);
-            product.Amount = request.Count;
+            cart.Items.Add(await _storageItemRepository.GetById(request.ProductId));
 
             await _cartRepository.Update(cart);
             await _cartRepository.SaveChangesAsync();
