+++ Application/Commands/CartCommands/UpdateProduct/UpdateProductInCartCommandHandler.cs
@@ -2,7 +2,6 @@
 using AutoMapper;
 using Data;
 using MediatR;
-using Microsoft.Extensions.Configuration;
 using Persistence.Repository;
 using System;
 using System.Collections.Generic;
@@ -10,28 +9,27 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Application.Commands.CartCommands.AddProduct
+namespace Application.Commands.CartCommands.UpdateProduct
 {
-    public class AddProductCommandHandler : IRequestHandler<AddProductCommand, CartModel>
+    public class UpdateProductInCartCommandHandler : IRequestHandler<UpdateProductInCartCommand, CartModel>
     {
         private readonly IRepository<Cart> _cartRepository;
-        private readonly IRepository<StorageItem> _storageItemRepository;
         private readonly IMapper _mapper;
 
-        public AddProductCommandHandler(IMapper mapper, IRepository<Cart> cartRepository,
-            IRepository<StorageItem> storageItemRepository)
+        public UpdateProductInCartCommandHandler(IMapper mapper, IRepository<Cart> cartRepository)
         {
             _cartRepository = cartRepository;
-            _storageItemRepository = storageItemRepository;
             _mapper = mapper;
         }
-        public async Task<CartModel> Handle(AddProductCommand request, CancellationToken cancellationToken)
+
+        public async Task<CartModel> Handle(UpdateProductInCartCommand request, CancellationToken cancellationToken)
         {
             if (request == null) throw new ArgumentException();
 
             var cart = await _cartRepository.GetById(request.CartId);
 
-            cart.Items.Add(await _storageItemRepository.GetById(request.ProductId));
+            var product = cart.Items.FirstOrDefault(item => item.Product.Id == request.ProductId);
+            product.Amount = request.Count;
 
             await _cartRepository.Update(cart);
             await _cartRepository.SaveChangesAsync();
